Change the positioning and use of Handle scopes
The v8 HandleScopes are now less forgiving when returning handles that
go out of the (handle) scope they were created in. You now have to
explicitly copy out such handles via EscapableHandleScope objects.

To make it all work, HandleScopes had to be moved off Runtime::Select
objects also. Instead requiring a handle scope to be set up when
needed for when a Runtime::Select context is entered.

Setting up HandleScopes for Function::Call() is also problematic, as
we do want its result to outlive the invocation. Remove it. (This also
addresses the debug asserts that have been triggering for a while.)