Add utility functions for handling constructors and prototypes
In base::Object, add GetPrototype()/SetPrototype(), for accessing and
manipulating an object's prototype.  Also add GlobalObject() for fetching
the current context's global object.

In base::Function, add Construct(), similar to Call() but for invoking
the function as a constructor.  Also remove an unimplemented and pointless
Call() overload.

In api::Class, add GetConstructor() and GetPrototype() for accessing the
class's constructor and prototype objects.  Also add an Inherit() overload
for inheriting an arbitrary object (setting the class's prototype object's
prototype to the argument object.)