Re-implement Bytes as a sub-class of Uint8Array
The v8::Object::SetIndexedPropertiesToExternalArrayData() method that we
previously used to implement it no longer works (it crashes when called,
to be specific) so something had to be done.  If V8 had had full built-in
support for typed arrays when I first implemented Bytes, I would have
preferred to implement Bytes (more or less) this way in the first place.