Detect HTTP request failure in URL.{get,put,post} methods
Parse the HTTP response status line, and throw a URLError if the
status code is not in the 2xx range.  The error thrown references
the Request object used internally, which contains all the
information about the actual response, for the caller to inspect.

Also expose the parsed HTTP status code as Request.statusCode on
all requests.