Support requesting a specific authentication method
Only "basic", "digest" and "any" are supported.  Doing this is sometimes
required to get authentication working.  One such case is when the server
replies with an authentication challenge containing an unsupported method,
but accepts e.g. basic authentication, in which case libcurl must be told
specifically to do basic authentication, or it won't try to authenticate
at all.