Fix incorrect formatting of SQL queries
An assumption was made that the return value of std::string::c_str() from
string objects in a std::vector<std::string> stayed valid as more string
objects where added to the vector. This is not a valid assumption, and it
turns out to be false in some environment.

Fix by using a std::list<std::string> to collect string objects instead.